% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bdm.R
\name{bdm}
\alias{bdm}
\title{Run one interaction of a Gillespie Algorithm of birth death and migration process in a system of generalized Lotka-Volterra system of competing species}
\usage{
bdm(N0, alphas, K, d0 = 0, b, m)
}
\arguments{
\item{N0}{vector of initial abundances of species in the community (set species not present to zero)}

\item{alphas}{matrix of interaction coefficients}

\item{K}{carrying capacities of species}

\item{d0}{death rate when N=0}

\item{b}{birth rates (constant)}

\item{m}{per capita migration rate in the metacommunity}
}
\description{
Run one interaction of a Gillespie Algorithm of birth death and migration process in a system of generalized Lotka-Volterra system of competing species
}

