% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Community.R
\name{Community}
\alias{Community}
\title{Rcpp module: Community}
\description{
Bla bla
}
\details{
\itemize{
 \item{multilayer}{ reflection and transmission of a multilayer using a transfer matrix formalism}
 \item{recursive <- fresnel}{ reflection and transmission of a multilayer using recursive application of Fresnel coefficients}
}
}

