% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{stoch_environment}
\alias{stoch_environment}
\title{Random environment generator}
\usage{
stoch_environment(time, sd, intervals = 10000)
}
\arguments{
\item{time}{Final time}

\item{sd}{Standard deviation of the random walk process}

\item{intervals}{Number of steps of the random walk}
}
\description{
This function is a simple generator for a stochastic data frame
that can be used to represent a randomly fluctuating environment
for use in \code{\link{Init_Community}}.
}

