% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{GillesComToFile}
\alias{GillesComFromFile}
\alias{GillesComToFile}
\alias{SeedFromFile}
\title{File persistence of simulations}
\usage{
GillesComToFile(community, file = "GillesCom.rda")

GillesComFromFile(file = "GillesCom.rda")

SeedFromFile(file = "GillesCom.rda")
}
\arguments{
\item{community}{An object of class \code{\link{Community}}}

\item{file}{filename to which save / from which load the data}
}
\description{
Functions to save and load simulated data from the disk
}
\details{
The function \code{GillesComToFile} saves all of the relevant
data about a simulated community to a file, including the random
number generator seed. This allows for retaking the simulation later
or even in other computers with reproducible results.

The function \code{GillesComFromFile} loads all of the relevant
data about a simulated community from a file. If you also want to
restore the number generator seed, use ".Random.seed <- SeedFromFile()".
}
\examples{
filename = tempfile()
# Generate a new community
a = Init_Community(50); bdm(a, 10000, progress="none")
# Saves the simulated data
GillesComToFile(a, filename)
# Loads the simulated data
b = GillesComFromFile(filename)
# Notice that a and b are different objects...
identical (a,b)
# But their content is the same
identical (a$trajectories, b$trajectories)
# To load the random seed as well...
.Random.seed <- SeedFromFile(filename)
}

