% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Community.R
\name{interaction}
\alias{interaction}
\alias{ones}
\title{Interaction matrix}
\usage{
interaction(J, stren = 0.1, con = 1, comp = TRUE)

ones(J)
}
\arguments{
\item{J}{size of metacommunity}

\item{stren}{strength of interaction matrix, which is the standard deviation of the Gaussin from which the values are drawn}

\item{con}{connectance of the interaction matrix. Defaults to 1 (totally connected)}

\item{comp}{Logical. Use \code{TRUE} for a competition only matrix (all entries are positive); \code{FALSE} for otherwise}
}
\description{
The function \code{interaction} generates a Lotka-Volterra interaction matrix, following May. For a 
Caswell matrix, use \code{diag(J)}, and for a Hubbell matrix, use \code{ones(J)}.
}

