% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bdm.R
\name{ls.m}
\alias{ls.m}
\title{Generates migration rates from a log-series metacommunity}
\usage{
ls.m(J, alpha, m)
}
\arguments{
\item{J}{size of metacommunity}

\item{alpha}{Fisher's alpha}

\item{m}{per species migration rate}
}
\description{
Generates migration rates from a log-series metacommunity
}

