% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Community.R
\name{ls_migration}
\alias{ls_migration}
\title{Helper functions}
\usage{
ls_migration(J, S, alpha, m)
}
\arguments{
\item{J}{expected size of metacommunity (total number of individuals)}

\item{S}{Expected number of species in the metacommunity}

\item{alpha}{Fisher's alpha of the metacommunity}

\item{m}{per species migration rate}
}
\description{
Generates migration rates from a log-series metacommunity. The user is expected
to provide either S or alpha, but not both.
}

