% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{diagPlots}
\alias{diagPlots}
\alias{octavOverTime}
\alias{radOverTime}
\title{Plotting functions}
\usage{
diagPlots(which = 1:4)

radOverTime(steps)

octavOverTime(steps, prop = TRUE)
}
\arguments{
\item{which}{numerical, may be a single number or a vector: which plots should be done (see Details)}

\item{steps}{number of intervals in which to cut the history}

\item{prop}{Logical. Should the octav be plotted using proportions (as opposed to absolute numbers)?}
}
\description{
These functions plot relevant information about a simulation
}
\details{
The function \code{diagPlots} plots up to four different diagnostic plots (controlled by the
parameter "which":
1 - Species richness over time
2 - Total individuals over time
3 - Number of species equivalents calculated from Shannon information index
4 - Kolmogorov-Smirnoff statistic between SADs at each time compared to the SAD at the last time
5 - Estimated Fisher's alpha over time (from a log-series fit)
6 - Estimated sdlog over time (from a lognormal fit)

The functions \code{radOverTime} and \code{octavOverTime} provide a superimposing plot
with the rad and octav, respectively, at distinct points in time.
}

