% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Community.R
\name{Community}
\alias{Community}
\title{Community}
\usage{
Community()
}
\description{
Functions for generating and altering \code{Community} objects
}
\details{
The \code{Community} object is one of the fundamental objects in the GillesCom package.
}

