% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Community.R
\name{Init_Community}
\alias{Init_Community}
\alias{abundance}
\alias{bdm}
\alias{time}
\title{Community}
\usage{
Init_Community(abundance, interaction, K = 1000, b = 1, m = 0.1, d0 = 0)

bdm()

abundance()

time()
}
\arguments{
\item{abundance}{vector of initial abundances of species in the community (set species not present to zero). For convenience, a single number is expanded as \code{rep(1,N)}.}

\item{interaction}{matrix of interaction coefficients, see \code{\link{Interaction}}.}

\item{K}{carrying capacities of each species}

\item{b}{birth rates (constant)}

\item{m}{per capita migration rate in the metacommunity}

\item{d0}{death rate when N=0}
}
\description{
Functions for generating and altering the simulated Community. Function \code{Init_Community}
initializes a simulation.

Function \code{bdm} runs one interaction of a Gillespie Algorithm of birth death and migration process in
a system of generalized Lotka-Volterra system of competing species

Function \code{abundance} returns the current abundance vector for the community.

Function \code{time} returns the current simulation time for the community (TBI!)
}
\details{
Because of the way the interaction with the underlying c++ code is implemented, only one
community may be simulated at a time. Calling \code{Init_Community} more than once will
overwrite the previous simulation objects!
}
\examples{
Init_Community(100)
for (i in 1:50000) bdm()
(ab <- as.numeric(abundance()))
if(require(sads)) {
   f <- fitlnorm(ab[ab>0])
   plot(f)
}
}

