% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file.R
\name{GillesComToFile}
\alias{GillesComFromFile}
\alias{GillesComToFile}
\title{File persistence of simulations}
\usage{
GillesComToFile(file = "GillesCom.rda")

GillesComFromFile(file = "GillesCom.rda")
}
\description{
Functions to save and load simulated data from the disk
}
\details{
The function \code{GillesComToFile} saves all of the relevant
data about a simulated community to a file, including the random
number generator seed. This allows for retaking the simulation later
or even in other computers with reproducible results.

The function \code{GillesComFromFile} loads all of the relevant
data about a simulated community from a file. It returns the saved random
number generator seed as an invisible vector, so if you want to continue the
simulation with the saved key, use ".Random.seed <- GillesComFromFile()".
}

