% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Community.R
\name{Init_Community}
\alias{Init_Community}
\alias{bdm}
\title{Community generating functions}
\usage{
Init_Community(abundance, interaction, K = 1000, b = 1, m = 0.1, d0 = 0,
  save.int = 1, stochastic = data.frame())

bdm(community, count, time, progress = c("text", "none"))
}
\arguments{
\item{abundance}{vector of initial abundances of species in the community (set species not present to zero). 
For convenience, a single number is expanded as \code{rep(1,N)} (this also happens with parameters K, d0, b and m).}

\item{interaction}{matrix of interaction coefficients, see \code{\link{interaction_matrix}}.}

\item{K}{vector with carrying capacities of each species}

\item{b}{vector with birth rates (constant)}

\item{m}{vector with per capita migration rate in the metacommunity. May be the given as the 
resulting list of the \code{\link{ls_migration}} function}

\item{d0}{vector with death rate when N=0}

\item{save.int}{History saving interval (in simulated time units)}

\item{stochastic}{Optional; only use if you want to include demographic stocasticity. A data.frame consisting of 
two columns. The first represents the time intervals of the stochastic variation, and the second represents the
multiplier to birth rates at that time intervals. See the vignettes for a more in-depth explanation.}

\item{community}{An object of class \code{\link{Community}}}

\item{count}{Number of cycles to be simulated}

\item{time}{Total time to be simulated (i.e., if the simulation clock is in t=10, bdm(time=10) will simulate up to t=20)}

\item{progress}{Should a text bar be used? Currently, "text" will produce a text based bar, and \code{NULL} will produce none.}
}
\value{
\code{Init_Community} returns an object of class \code{\link{Community}}. 
\code{bdm} returns nothing. This function is called by its side effects only.
}
\description{
Functions for generating and altering the simulated Community. Function \code{Init_Community}
initializes a simulation.

Function \code{bdm} runs one interaction of a Gillespie Algorithm of birth death and migration process in 
a system of generalized Lotka-Volterra system of competing species. You may provide the number of simulated
cycles or the end time for the simulation, but not both.
}
\details{
Because of the way the interaction with the underlying c++ code is implemented, saving a Community
object using save() or exiting R will not work! See \code{\link{GillesComToFile}} for an alternative.
}
\examples{
# Initializes the community
Com = Init_Community(100)
# Runs some iteractions of the birth-death-migration process
bdm(Com, 5e5, progress = "none")
# Gets and analyzes the abundance vector
print(ab <- as.numeric(Com$abundance))
require(sads)
f <- sads::fitlnorm(ab[ab>0])
plot(f, which=1)
# Simulation internal time elapsed
print(Com$time)
# Trajectories saves a line for each time period elapsed (starting with 0), along with 
# the exact time at each snapshot:
print(length(Com$trajectories[[2]]))
}

